fixing ambiguity between command redirection and comparison operators
