+++ parser/src/aspects.rs
@@ -1,6 +1,7 @@
-pub(super) mod binary_operations_parser;
 pub(super) mod call_parser;
 pub(super) mod group_parser;
 pub(super) mod literal_parser;
 pub(super) mod var_declaration_parser;
 pub(super) mod var_reference_parser;
+pub(super) mod redirection_parser;
+pub(super) mod binary_operation_parser;
